Earthbound Game Recreation for Comp and Media II
I remade the classic Nintendo JRPG game Earthbound in Javascript. It is a bite-sized experience for people new to the series and the genre. It has been created with the intention of being a remix of creative and original game assets. 

The game was to be more ambitious than the final result, however, my IDE stopped allowed me to upload assets so this is now the final product.